+++ pypsa/components/legacy.py
@@ -12,6 +12,7 @@ from pypsa.common import UnexpectedError
 from pypsa.components._types import (
     Buses,
     Carriers,
+    FlowBasedConstraints,
     Generators,
     GlobalConstraints,
     Lines,
@@ -42,6 +43,7 @@ if TYPE_CHECKING:
 _CLASS_MAPPING = {
     "Bus": Buses,
     "Carrier": Carriers,
+    "FlowBasedConstraint": FlowBasedConstraints,
     "Generator": Generators,
     "GlobalConstraint": GlobalConstraints,
     "Line": Lines,